Failed to introduce chunk-based pipe storage

+++ src/main/java/com/neep/neepmeat/block/TankBlock.java
@@ -36,7 +36,7 @@ public class TankBlock extends BaseColumnBlock implements BlockEntityProvider
         {
             if (world.getBlockEntity(pos) instanceof TankBlockEntity be)
             {
-                player.sendMessage(Text.of(Long.toString(be.getBuffer(null).getAmount() / FluidConstants.BUCKET)), true);
+                player.sendMessage(Text.of(Float.toString(be.getBuffer(null).getAmount() / (float) FluidConstants.BUCKET)), true);
             }
         }
         return ActionResult.SUCCESS;
