+++ src/main/java/com/neep/neepmeat/block/machine/FluidDrainBlock.java
@@ -40,7 +40,7 @@ public class FluidDrainBlock extends BaseBlock implements BlockEntityProvider
         {
             if (world.getBlockEntity(pos) instanceof TankBlockEntity be)
             {
-                player.sendMessage(Text.of(Long.toString(be.getBuffer(null).getAmount() / FluidConstants.BUCKET)), true);
+                player.sendMessage(Text.of(Float.toString(be.getBuffer(null).getAmount() / (float) FluidConstants.BUCKET)), true);
             }
         }
         return ActionResult.SUCCESS;
